requirements.txt
tensorflow==2.12.0
tensorflow_hub==0.13.0
pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.4.0
requests>=2.28.0
